Move PropertyChange struct to CopyUtils namespace
