+++ src/CopyUtils/PropertyChange.cs.meta
