+++ test/Point.cs
@@ -1,4 +1,5 @@
 using System;
+using AReSSO.CopyUtils;
 
 namespace AReSSO.Test
 {
