+++ usageExamples/ExamplePerson.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using AReSSO.CopyUtils;
 
 namespace AReSSO.UsageExamples
 {
